+++ odb/source-packed.c
@@ -503,7 +503,7 @@ static int odb_source_packed_freshen_object(struct odb_source *source,
 
 static int odb_source_packed_write_object(struct odb_source *source UNUSED,
 					  const void *buf UNUSED,
-					  unsigned long len UNUSED,
+					  size_t len UNUSED,
 					  enum object_type type UNUSED,
 					  struct object_id *oid UNUSED,
 					  struct object_id *compat_oid UNUSED,
